pyproject.toml
[tool.poetry]
name = "west"
version = "0.1.0"
description = "Send notifications to client side via websockets"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^13.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
setuptools = "^75.2.0"
redis = "^5.1.1"
typer = "^0.12.5"
pyyaml = "^6.0.2"
pyjwt = "^2.9.0"

[tool.taskipy.tasks]
server = "python west/main.py"
settings = "python west/cli/main.py"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
